+++ source/extensions/omni.isaac.benchmarks/omni/isaac/benchmarks/tests/test_benchmark_physx_lidar.py
@@ -77,5 +77,11 @@ class TestBenchmarkLidar(BaseIsaacBenchmark):
     async def test_benchmark_10_physx_lidar(self):
         await self.benchmark_physx_lidar(10)
 
+    async def test_benchmark_25_physx_lidar(self):
+        await self.benchmark_physx_lidar(25)
+
     async def test_benchmark_50_physx_lidar(self):
         await self.benchmark_physx_lidar(50)
+
+    async def test_benchmark_100_physx_lidar(self):
+        await self.benchmark_physx_lidar(100)